Store canvas rect
