+++ src/engine/editor.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <glm/vec2.hpp>
+#include <platform/rect.h>
 
 namespace platform {
 	struct Input;
@@ -14,10 +15,10 @@ namespace engine {
 	struct EditorState {
 		int zoom_index;
 		float zoom;
-		glm::vec2 canvas_pos;
+		platform::Rect canvas;
 	};
 
-	EditorState init_editor();
+	EditorState init_editor(glm::vec2 canvas_size);
 	void update_editor(EditorState* editor, const platform::Input* input, platform::PlatformAPI* platform);
 	void render_editor(platform::Renderer* renderer, const EditorState* editor, platform::Canvas canvas);
