+++ src/platform/renderer.h
@@ -2,6 +2,7 @@
 
 #include <platform/font.h>
 #include <platform/image.h>
+#include <platform/rect.h>
 #include <platform/texture.h>
 
 #include <SDL2/SDL.h>
@@ -36,26 +37,6 @@ namespace platform {
 		GLuint id;
 		GLuint vao;
 		GLuint vbo;
-	};
-
-	// Origin in upper left corner (e.g. xy-coordinates)
-	// top_left o-----------o
-	//          |           |
-	//          |           |
-	//          o-----------o bottom_right
-	struct Rect {
-		glm::vec2 top_left;
-		glm::vec2 bottom_right;
-	};
-
-	// Origin in bottom left corner (e.g. uv-coordinates)
-	//             o-----------o top_right
-	//             |           |
-	//             |           |
-	// bottom_left o-----------o
-	struct FlipRect {
-		glm::vec2 bottom_left;
-		glm::vec2 top_right;
 	};
 
 	enum class ShaderProgramError {
